feat(auth): add ForwardAuth support via X-Plex-Token header
This adds ForwardAuth support based on a user's Plex token being sent as an
HTTP header (X-Plex-Token) within a request to Overseerr. When a session has
no current user, but this feature is enabled and the header is present,
Overseerr will attempt to create and/or log in the user in a manner nearly
identical to the /api/v1/auth/plex endpoint.